@@ -3,9 +3,11 @@ package com.neutobo.recyclerviewwithads
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.GridLayoutManager
-import kotlinx.android.synthetic.main.activity_main.*
+import com.neutobo.recyclerviewwithads.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
 
     private var adapter: ThingAdapterWithAds = ThingAdapterWithAds()
     private var repository: FakeDataRepository = FakeDataRepository()
@@ -13,11 +15,13 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
         layoutManager = GridLayoutManager(this, 2)
 
-        mainRecyclerView.layoutManager = layoutManager
-        mainRecyclerView.adapter = adapter
+        binding.mainRecyclerView.layoutManager = layoutManager
+        binding.mainRecyclerView.adapter = adapter
 
         layoutManager.spanSizeLookup = object : GridLayoutManager.SpanSizeLookup() {
             override fun getSpanSize(position: Int): Int {
